deleted the legacyBehavior in the StyledLink
