--- lib/db/routes.json
@@ -1,46 +1,0 @@
-[
-  {
-    "id": "0",
-    "name": "The Five Lakes Trail to Fernsteinsee",
-    "activity": "trailrunning",
-    "difficulty": "moderate",
-    "length": "10 km",
-    "altitude": "1307 hm",
-
-    "description": "Great for any fitness level. Easily-accessible paths. Suitable for all skill levels. The starting point of the Tour is accessible with public transport.",
-    "imageUrl": "https://images.unsplash.com/photo-1662848732896-7b583983f997?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1474&q=80"
-  },
-  {
-    "id": "1",
-    "name": "Kranzhorn",
-    "activity": "hiking",
-    "difficulty": "easy",
-    "length": "9.2 km",
-    "altitude": "888 hm",
-
-    "description": "The Kranzhorn (1365m) in the Inntal valley is one of the most famous peaks thanks to its two summit crosses (the border runs exactly on the summit). It is easy to climb, and the Kranzhornhütte (Kranzhorn Shelter) just below the summit is a great place to rest and refresh. Here, children can also enjoy themselves with a playground and animals. Despite its relatively low elevation, the view is impressive: The Inntal valley stretches out at your feet, and the panorama extends from Heuberg to Wendelstein and Pendling, all the way to the Großglockner.",
-    "imageUrl": "https://images.unsplash.com/photo-1609924816440-5b7f56d14205?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1470&q=80"
-  },
-  {
-    "id": "2",
-    "name": "Jubiläumsgrat",
-    "activity": "climbing",
-    "difficulty": "difficult",
-    "length": "24.9 km",
-    "altitude": "2299 hm",
-
-    "description": "The route across the Jubiläumsgrat to the Zugspitze or from there to the Alpspitze is an absolute classic in the Bavarian Alps. However, the route should not be considered solely as a via ferrata, as only a few sections (especially in the stretch from the Biwakschachtel to the Alpspitze) are secured with steel cables. Primarily, the Jubiläumsgrat is an unsecured ridge scramble in the II-III difficulty level (partly extremely exposed). Sure-footedness, a head for heights, and good physical condition are essential prerequisites for this tour. From the Zugspitze to the Alpspitze, the Jubiläumsgrat mostly stays above 2600m, so favorable and stable weather conditions are crucial here.",
-    "imageUrl": "https://images.unsplash.com/photo-1662446325301-8e4970b85dcc?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1470&q=80"
-  },
-  {
-    "id": "3",
-    "name": "Großglockner Normalweg",
-    "activity": "hiking",
-    "difficulty": "difficult",
-    "length": "8 km",
-    "altitude": "1737 hm",
-
-    "description": "The Großglockner, standing at a height of 3798 meters, is the highest mountain in Austria. The normal route offers a diverse alpine experience, including glacier crossings, easy rock climbing, and a slightly exposed summit ridge. From the summit, you can enjoy breathtaking views of the Ortler to the west, the Watzmann to the north, and the Dolomites to the south. It's a remarkable location to be surrounded by such magnificent peaks. During the ascent, we will spend the night at the Erzherzog Johann Hütte, also known as Adlersruhe, situated at 3454 meters. This hut has been open since 1880 and serves as a convenient stop along the way to the Großglockner, providing a place for rest and refreshments.",
-    "imageUrl": "https://images.unsplash.com/photo-1630709555838-3ccc2a8fb7ac?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1548&q=80"
-  }
-]
